moved cli actions into their own folder, added server action
