+++ src/khamake.ts
@@ -12,7 +12,7 @@ import { Options } from './Options';
 import { Platform } from './Platform';
 import { VisualStudioVersion } from './VisualStudioVersion';
 
-import { CommandLineAction, CommandLineFlagParameter, CommandLineParser, CommandLineParameter } from '@microsoft/ts-command-line';
+import { CommandLineParser } from '@microsoft/ts-command-line';
 
 let defaultTarget: string;
 if (os.platform() === 'linux') {
@@ -368,47 +368,8 @@ else {
 }
 */
 
-class InitAction extends CommandLineAction {
-	private _name: CommandLineParameter;
-	private _from: CommandLineParameter;
-	private _projectFile: CommandLineParameter;
-
-	public constructor() {
-		super({
-			actionName: 'init',
-			summary: 'Init a Kha project',
-			documentation: 'Init a Kha project inside the current directory'
-		});
-	}
-
-	protected onExecute(): Promise<void> { // abstract
-		// TODO: actually make it run!
-		return Promise.resolve();
-	}
-
-	protected onDefineParameters(): void { // abstract
-		this._name = this.defineStringParameter({
-			argumentName: "NAME",
-			parameterShortName: "-n",
-			parameterLongName: "--name",
-			description: "Project name to use when initializing a project",
-			defaultValue: "Project"
-		});
-		this._from = this.defineStringParameter({
-			argumentName: "PATH",
-			parameterShortName: "-f",
-			parameterLongName: "--from",
-			description: "Location of your project",
-			defaultValue: "."
-		});
-		this._projectFile = this.defineStringParameter({
-			argumentName: "PATH",
-			parameterLongName: "--projectfile",
-			description: "Name of your project file",
-			defaultValue: "khafile.js"
-		});
-	}
-}
+import { InitAction } from './cli/InitAction';
+import { ServerAction } from './cli/ServerAction';
 
 class KhamakeCommandLine extends CommandLineParser {
 	constructor() {
@@ -425,6 +386,7 @@ class KhamakeCommandLine extends CommandLineParser {
 
 	private _populateActions(): void {
 		this.addAction(new InitAction());
+		this.addAction(new ServerAction());
 	}
 }
